pyproject.toml
[project]
name = "nonebot-plugin-omikuji"
version = "0.1.0"
description = "Nonebot2的御神签插件"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.3",
    "nonebot-plugin-suggarchat>=3.4.4.2",
    "nonebot-adapter-onebot>=2.4.6",
]
license = { text = "GPL-3.0-or-later" }
keywords = ['chat", "suggar", "nonebot']


[project.urls]
"Homepage" = "https://github.com/LiteSuggarDEV/nonebot_plugin_omikuji"
"Source" = "https://github.com/LiteSuggarDEV/nonebot_plugin_omikuji"
"Issue Tracker" = "https://github.com/LiteSuggarDEV/nonebot_plugin_omikuji/issues"

[tool.uv]
dev-dependencies = [
    "ruff>=0.12.8",
    "nonebot-plugin-orm[default]>=0.8.2",
    "pip>=25.2",
]


[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
    "I",     # isort
    "PERF",  # pylint-performance
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "E501",   # line-too-long
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.pyright]
typeCheckingMode = "standard"

[tool.nonebot]
plugins = ["nonebot_plugin_suggarchat", "nonebot_plugin_localstore","nonebot_plugin_omikuji"]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
